+++ src/scenes/gameScene.ts
@@ -52,6 +52,8 @@ export class GameScene extends Scene {
             ['green-portal', 'portalgreen.png'],
             ['ice', 'iceblock.png'],
             ['vine', 'vines.png'],
+            ['mud-block', 'mudblock.png'],
+            ['no-jump-block', 'nojumpblock.png'],
         ];
 
         for (let i = 0; i < 10; i++) // adding all 10 doors/keys in with code cause im lazy!!!